WL, Inference Time, AUROC, TNR
For training models of different window lengths change the wl parameter
in training and inference script. The corresponding value of input
dimension to layer 4 in the model also has to be changed. Easiest way is
to run and find out what is the expected dimension. Ideally, for
different window size, you should customise model.py as well.